feat(config): add formatting policies

+++ internal/configgen/testdata/badtype/patch.go
@@ -2,5 +2,5 @@ package fixture
 
 type Patch struct {
 	// Retries is a type the generator has no mapping for.
-	Retries *int `json:"retries"`
+	Retries *float64 `json:"retries"`
 }
